requirements.txt
asgiref==3.5.2
Django==4.1
django-filter==23.3
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.5
packaging==23.2
pip==23.3
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg-pool==3.1.8
psycopg2-binary==2.9.9
pytz==2023.3.post1
setuptools==68.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
wheel==0.41.2
django-bootstrap-v5
slack_sdk
